use int for elevation

+++ sport-log-client/lib/models/cardio/position.dart
@@ -17,7 +17,7 @@ class Position {
 
   @JsonKey(name: "lo") double longitude;
   @JsonKey(name: "la") double latitude;
-  @JsonKey(name: "e") double elevation;
+  @JsonKey(name: "e") int elevation;
   @JsonKey(name: "d") int distance;
   @JsonKey(name: "t") int time;
 
@@ -30,7 +30,7 @@ class Position {
     final bytes = ByteData(byteSize)
       ..setFloat64(0, longitude)
       ..setFloat64(8, latitude)
-      ..setFloat64(16, elevation)
+      ..setInt64(16, elevation)
       ..setInt64(24, distance)
       ..setInt64(32, time);
     final list = bytes.buffer.asUint8List();
@@ -44,7 +44,7 @@ class Position {
     return Position(
       longitude: bytes.getFloat64(0),
       latitude: bytes.getFloat64(8),
-      elevation: bytes.getFloat64(16),
+      elevation: bytes.getInt64(16),
       distance: bytes.getInt64(24),
       time: bytes.getInt64(32),
     );
